requirements.txt
asgiref==3.2.7
autopep8==1.5.3
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.3.0
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.1.2
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
